LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/request-upload.any-expected.txt

CONSOLE MESSAGE: line 27: ReferenceError: Can't find variable: RESOURCES_DIR

Harness Error (FAIL), message = ReferenceError: Can't find variable: RESOURCES_DIR
